Change NXPower Builds to use Centos 7.7 instead of Centos 7.3
Centos 7.3.1611 no longer builds with EPEL 7 as there are unresolved
package dependencies with EPEL 7.  The package python-srpm-macros and
a couple more were introduced in Centos 7.7.  EPEL does not provide
point in time snapshots of its releases, so the only resolution to
this problem is to upgrade the distro to Centos 7.7.

This drives a second set of changes in that some of the packages
provided by NXPower are now downlevel to the distro.  None of these
packages were modified by Nutanix, so they are simply being removed
from the build list.  The distro provided packages will be used instead.
Downlevel packages to the distro that are being removed:

- gcc                       Old version: 4.8.5-12   Distro: 4.8.5-39
- librtas                   Old version: 1.4.1      Distro: 2.0.1-2
- libservicelog             Old version: 1.1.16-2   Distro: 1.1.18-1

Signed-off-by: Luke Browning [IBM] <918375@nutanix.com>