Change NXPower Builds to use Centos 7.7 instead of Centos 7.3
Centos 7.3.1611 no longer builds with EPEL 7 as there are unresolved
package dependencies.  EPEL 7 has a dependency on the package
python-srpm-macros which was introduced in Centos 7.7.  There are a
couple of other missing dependencies as well.  EPEL does not provide
point in time snapshots of its code, so the only resolution is to
upgrade the distro to Centos 7.7.

This drives a second set of changes in that some of the packages
provided by NXPower are now downlevel to the distro.  None of these
packages were modified by Nutanix, so they are simply being removed
from the build list.  The distro provided packages will be used instead.
Downlevel packages to the distro that are being removed:

- gcc                   Old version: 4.8.5-12     Distro: 4.8.5-39
- librtas               Old version: 1.4.1        Distro: 2.0.1-2
- libservicelog         Old version: 1.1.16-2     Distro: 1.1.18-1
- crash                 Old version: 7.1.6.1      Distro: 7.2.3.10
- SLOF                  Old version: 20170303-2   Distro: 20171214-3
- libvpd                Old version: 2.2.5-4      Distro: 2.2.6
- lsvpd                 Old version: 1.7.7-6      Distro: 1.7.9-1

There are also directories in NXPower/versions for projects such as
docker, docker-swarm, flannel, and golang.  These directories are not
built by NXPower/builds.  They are simply inherited from the cloned
project open-power-host-os/versions.  They could be removed.

The list of projects to be built are identified in the file
NXPower/versions/open-power-host-os/open-power-host-os.yaml.  After
removing the obsolete packages above, only five projects remain:
hwdata, kernel, libvirt, qemu, and servicelog.  openvswitch is also
built as it is specified in the %build section of the spec file
open-power-host-os.spec.  Finally, the package nutanix-ahv and its
dependent package tunctl are built as they are explicitly listed as
package requirements in NXPower/builds/config.yaml.

Signed-off-by: Luke Browning [IBM] <918375@nutanix.com>